Created method DivideByBigInteger. Changed ctor parameter serviceTime (now value has to be in seconds)
In some case we can't divide our numbers because to divide some number by second number, the divisor should be as one of the system data type and that's why when we try to cast BigInteger to ulong sometimes we can get exception. In method DivideByBigInteger() firstly we try to cast out number to ulong data type. If we can't do it we catch exception and in this way we divide our divided by ulong max value and our BigInteger divisor by ulong max value. Try to cast it again. And when we finally can successfully cast our divisor to ulong we divide divided number by casted divisor and get result.